requirements.txt
websocket_client==1.4.2
rel==0.4.7
python-dotenv==0.21.0
pymongo==4.3.3